Move creation of AppDatabase to StorageModule
