+++ app/src/androidTest/java/org/simple/clinic/TestClinicApp.kt
@@ -69,10 +69,10 @@ class TestClinicApp : ClinicApp() {
     // We have moved the in-memory database configuration to the sqlite openhelper factory
     // but we still have to provide a non-empty name for Room, otherwise it complains.
     return DaggerTestAppComponent.builder()
-        .appModule(object : AppModule(this, "ignored-db-name", runDatabaseQueriesOnMainThread = true) {
+        .appModule(object : AppModule(this) {
           override fun clock(): Clock = TestClock()
         })
-        .storageModule(object : StorageModule() {
+        .storageModule(object : StorageModule(databaseName = "ignored", runDatabaseQueriesOnMainThread = true) {
           override fun sqliteOpenHelperFactory() = AppSqliteOpenHelperFactory(inMemory = true)
         })
         .syncModule(object : SyncModule() {
