+++ app/src/main/java/org/simple/clinic/storage/StorageModule.kt
@@ -2,15 +2,58 @@ package org.simple.clinic.storage
 
 import android.app.Application
 import android.arch.persistence.db.SupportSQLiteOpenHelper
+import android.arch.persistence.room.Room
 import android.content.SharedPreferences
 import android.preference.PreferenceManager
 import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
+import org.simple.clinic.AppDatabase
+import org.simple.clinic.di.AppScope
 import org.simple.clinic.di.AppSqliteOpenHelperFactory
 
 @Module
-open class StorageModule {
+open class StorageModule(
+    private val databaseName: String = "red-db",
+    private val runDatabaseQueriesOnMainThread: Boolean = false
+) {
+
+  @Provides
+  @AppScope
+  fun appDatabase(appContext: Application, factory: SupportSQLiteOpenHelper.Factory): AppDatabase {
+    return Room.databaseBuilder(appContext, AppDatabase::class.java, databaseName)
+        .openHelperFactory(factory)
+        .apply {
+          if (runDatabaseQueriesOnMainThread) {
+            allowMainThreadQueries()
+          }
+        }
+        .addMigrations(
+            Migration_3_4(),
+            Migration_4_5(),
+            Migration_5_6(),
+            Migration_6_7(),
+            Migration_7_8(),
+            Migration_8_9(),
+            Migration_9_10(),
+            Migration_10_11(),
+            Migration_11_12(),
+            Migration_12_13(),
+            Migration_13_14(),
+            Migration_14_15(),
+            Migration_15_16(),
+            Migration_16_17(),
+            Migration_17_18(),
+            Migration_18_19(),
+            Migration_19_20(),
+            Migration_20_21(),
+            Migration_21_22(),
+            Migration_22_23(),
+            Migration_23_24(),
+            Migration_24_25(),
+            Migration_25_26())
+        .build()
+  }
 
   @Provides
   fun rxSharedPreferences(appContext: Application): RxSharedPreferences {
